refactor(worker): move GcodeAnalyser to worker
- Moved GcodeAnalyser to worker/utilities/gcode/
- Removed implicit dependency on GRBL constants
- Inlined GCODE_PROGRAM_END_CODES in gateway/fileExecutor.py
- Deleted core/utilities/gcode/ (no remaining consumers)
