+++ src/worker/worker/utilities/gcode/gcodeAnalyser.py
@@ -3,8 +3,6 @@ from collections import Counter
 from pathlib import Path
 from typing import Pattern
 
-from core.utilities.gcode.constants import GCODE_VALID_GCODES, GCODE_VALID_MCODES
-
 # Regular expressions to classify commands
 pause_pattern = re.compile(r"^(?:N\d+\s+)?M(0|1|00|01)\s*", re.MULTILINE)
 move_pattern = re.compile(r"^(?:N\d+\s+)?G(0|1|00|01)\s*", re.MULTILINE)
@@ -20,12 +18,17 @@ t_pattern = re.compile(r"^(?!;|\().*(T\d+)", re.MULTILINE)
 class GcodeAnalyser:
     """Utility class to get values of interest from a gcode file."""
 
-    def __init__(self, file_path: str | Path):
-        # Attributes definition
+    def __init__(
+        self,
+        file_path: str | Path,
+        valid_gcodes: list[str],
+        valid_mcodes: list[str],
+    ):
         self.file_path = file_path
+        self._valid_gcodes = valid_gcodes
+        self._valid_mcodes = valid_mcodes
 
     def analyse(self):
-        # Initialize values of interest
         total_lines = 0
         comment_count = 0
         movement_lines = 0
@@ -36,26 +39,21 @@ class GcodeAnalyser:
         commands_mcode = {}
         unsupported_commands = []
 
-        # Static analysis
         with open(self.file_path, "r") as gcode:
             content = gcode.read()
 
             total_lines = len(content.splitlines())
-
             pause_count = self._count(pause_pattern, content)
             movement_lines = self._count(move_pattern, content)
             comment_count = self._count(comment_pattern, content)
-
             tools = self._find_all_unique(t_pattern, content)
-
             max_feedrate = self._find_max(feedrate_pattern, content)
-
             commands_gcode = self._count_all(gcode_pattern, content)
             commands_mcode = self._count_all(mcode_pattern, content)
 
             unsupported_commands = list(
-                filter(lambda x: x not in GCODE_VALID_GCODES, commands_gcode.keys())
-            ) + list(filter(lambda x: x not in GCODE_VALID_MCODES, commands_mcode.keys()))
+                filter(lambda x: x not in self._valid_gcodes, commands_gcode.keys())
+            ) + list(filter(lambda x: x not in self._valid_mcodes, commands_mcode.keys()))
 
         return {
             "total_lines": total_lines,
@@ -71,18 +69,15 @@ class GcodeAnalyser:
     # UTILITIES
 
     def _count(self, regex: Pattern, text: str):
-        matches = regex.findall(text)
-        return len(matches)
+        return len(regex.findall(text))
 
     def _count_all(self, regex: Pattern, text: str):
-        matches = regex.findall(text)
-        return dict(Counter(matches))
+        return dict(Counter(regex.findall(text)))
 
     def _find_all_unique(self, regex: Pattern, text: str):
         matches = regex.findall(text)
         if not matches:
             return []
-
         list_set = list(set(matches))
         list_set.sort()
         return list_set
@@ -91,6 +86,4 @@ class GcodeAnalyser:
         matches = regex.findall(text)
         if not matches:
             return 0
-
-        values = [int(x) for x in matches]
-        return max(values)
+        return max(int(x) for x in matches)
